backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
azure-storage-blob==12.19.0
python-dotenv==1.0.0
bcrypt==4.0.1
pydantic[email]
PyJWT==2.8.0